+++ unsupported_plugins/toasim/makeRedNoise.h
@@ -3,22 +3,22 @@
 
 
 typedef struct rednoisemodel {
-	float start;   // start MJD
-	float end;     // end MJD
+	double start;   // start MJD
+	double end;     // end MJD
 	int npt;       // points per realisation
 	int nreal;     // number of realisations
-	float pwr_1yr; // power at 1 year
-	float index;   // index in power spectrum
-	float cutoff;  // model is zero below 'cutoff' (yr^-1)
-	float flatten; // model is flat below 'flatten'(yr^-1)
-	float* data;   // data
-	float tres;
+	double pwr_1yr; // power at 1 year
+	double index;   // index in power spectrum
+	double cutoff;  // model is zero below 'cutoff' (yr^-1)
+	double flatten; // model is flat below 'flatten'(yr^-1)
+	double* data;   // data
+	double tres;
 	char mode;
 } rednoisemodel_t;
 
 
-rednoisemodel_t* setupRedNoiseModel(float start,float end, int npt, int nreal, float amp_1yr, float index);
+rednoisemodel_t* setupRedNoiseModel(double start,double end, int npt, int nreal, double amp_1yr, double index);
 void populateRedNoiseModel(rednoisemodel_t* model,long seed);
-float getRedNoiseValue(rednoisemodel_t* model, float mjd,int real);
+double getRedNoiseValue(rednoisemodel_t* model, double mjd,int real);
 void freeRedNoiseModel(rednoisemodel_t* model);
-float* getPowerSpectrum(rednoisemodel_t* model);
+double* getPowerSpectrum(rednoisemodel_t* model);
